Extract Main and applicationHref from careers
