+++ src/components/careers/applicationHref.js
@@ -1,0 +1,2 @@
+export default (typeformSlug) =>
+  `https://roadiehq.typeform.com/to/cdF3Ls?roleslug=${typeformSlug}&utm_source=roadie.io`;
